Add files via upload
This is an Android App for Communication between Arduino devices and Android App. Android App will act as a smart switch in that case. Different person can control a particular device if it is registered or signed in to Android App.
These are files included for Android App which acts as a switch for different Arduino Devices Connected with it. 
The Android App includes following:
- MQTT Protocol Implementation for Communication
- Firebase Authentication Integration for Log in/Sign Up Admin Panels
- Firebase Database Integration
- Customized Adapter View